ci(python): install the project so the gauntlet Python gate really checks it
The Python job installed only ruff/mypy/pytest. Without the package and its
sibling toolkit, mypy could not resolve a single third-party import and
pytest could not have collected anything, so the gate was decorative. It now
fetches b2b-agent-toolkit the same way tests.yml does, installs this project,
and adds the WeasyPrint system libraries it loads at import time.

That turned on real checking, and these were the findings:

- ruff: datetime.UTC over the deprecated timezone.utc alias, sorted imports,
  a dict() call rewritten as a literal, and ruff format applied.
- pdf.py wrapped round() in int(); round() without ndigits already returns an
  int, so the cast was dead.
- nodes.py read .text straight off the first Anthropic content block. That
  block is a union of a dozen kinds and only one of them carries prose, so it
  now asserts a TextBlock and fails with a clear message otherwise.
- yaml has no inline types; added the types-PyYAML stub package to dev deps.
- WeasyPrint ships no annotations and has no typeshed stubs, so mypy is told
  to follow its untyped source (real inference) rather than treat it as Any.

+++ src/compliance/pdf.py
@@ -9,7 +9,9 @@ from jinja2 import Environment, FileSystemLoader, select_autoescape
 from weasyprint import HTML
 
 _TEMPLATES = Path(__file__).resolve().parents[2] / "templates"
-_env = Environment(loader=FileSystemLoader(_TEMPLATES), autoescape=select_autoescape(["html", "xml"]))
+_env = Environment(
+    loader=FileSystemLoader(_TEMPLATES), autoescape=select_autoescape(["html", "xml"])
+)
 
 
 def render_evidence_pack(
@@ -30,7 +32,7 @@ def render_evidence_pack(
         results=results,
         passed=passed,
         failed=failed,
-        pass_rate_pct=int(round(len(passed) / max(len(results), 1) * 100)),
+        pass_rate_pct=round(len(passed) / max(len(results), 1) * 100),
         summary_md=summary_md,
     )
     return HTML(string=html, base_url=str(_TEMPLATES)).write_pdf()
